refactor: extract printer to it's own file
